Refactor batcher
