--- pkg/batcher/server.go
@@ -1,32 +1,0 @@
-/*
-Copyright 2020 kubeflow.org.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package batcher
-
-import (
-	"github.com/astaxie/beego"
-	_ "github.com/kubeflow/kfserving/pkg/batcher/routers"
-	"github.com/kubeflow/kfserving/pkg/constants"
-	"strconv"
-)
-
-func StartHttpServer() {
-	beego.BConfig.AppName = "batcher"
-	beego.BConfig.Listen.HTTPPort, _ = strconv.Atoi(constants.InferenceServiceDefaultBatcherPort)
-	beego.BConfig.RunMode = "dev"
-	beego.BConfig.CopyRequestBody = true
-	beego.Run()
-}
